hotel-reservations/frontend-service/requirements.txt
Flask
requests
pyjwt
prometheus-client